+++ src/Profile/Magento/Gateway/Local/Reader/SalesChannelReader.php
@@ -21,6 +21,7 @@ abstract class SalesChannelReader extends AbstractReader
 
         $storeGroups = $this->mapData($this->fetchStoreGroups($migrationContext), [], ['storeGroup', 'website']);
         $groupIds = \array_column($storeGroups, 'group_id');
+        $websiteIds = \array_column($storeGroups, 'website_id');
 
         $storeViews = $this->mapData($this->fetchStoreViews($groupIds), [], ['storeView']);
         $storeIds = [];
@@ -42,15 +43,10 @@ abstract class SalesChannelReader extends AbstractReader
         $locales = $this->fetchLocales($storeIds);
 
         $defaults = $this->fetchDefaults();
+        $websiteConfigs = $this->fetchWebsiteConfig($websiteIds);
 
         foreach ($storeGroups as &$store) {
-            $store['currencies'] = $defaults['defaultAllowedCurrencies'];
-            $store['countries'] = $defaults['defaultAllowedCountries'];
-            $store['locales'][] = $defaults['defaultLocale'];
-
-            $store['defaultCurrency'] = $defaults['defaultCurrency'];
-            $store['defaultCountry'] = $defaults['defaultCountry'];
-            $store['defaultLocale'] = $defaults['defaultLocale'];
+            $this->setDefaultConfig($store, $defaults, $websiteConfigs);
 
             foreach ($store['storeViews'] as $storeView) {
                 $storeId = $storeView['store_id'];
@@ -72,8 +68,8 @@ abstract class SalesChannelReader extends AbstractReader
             }
             $store['locales'] = \array_unique($store['locales']);
             $store['currencies'] = \array_unique($store['currencies']);
-            $store['carriers'] = $carriers;
-            $store['payments'] = $payments;
+            $this->setCarriers($store, $carriers);
+            $this->setPayments($store, $payments);
         }
 
         return $this->cleanupResultSet($storeGroups);
@@ -201,47 +197,157 @@ SQL;
     protected function fetchCarriers(): array
     {
         $sql = <<<SQL
-SELECT carrier.*
+SELECT carrier_active.*
 FROM (
           SELECT
-            REPLACE(REPLACE(config.path, '/title', ''), 'carriers/', '') AS carrier_id,
-            config.*
+            REPLACE(REPLACE(config.path, '/title', ''), 'carriers/', '') AS carrier_id
           FROM {$this->tablePrefix}core_config_data config
           WHERE path LIKE 'carriers/%/title' AND scope = 'default'
       ) AS carrier,
       (
           SELECT
-            REPLACE(REPLACE(config.path, '/active', ''), 'carriers/', '') AS carrier_id
+            REPLACE(REPLACE(config.path, '/active', ''), 'carriers/', '') AS carrier_id,
+            config.*
           FROM {$this->tablePrefix}core_config_data config
-          WHERE path LIKE 'carriers/%/active' AND scope = 'default' AND value = true
+          WHERE path LIKE 'carriers/%/active'
       ) AS carrier_active
-WHERE carrier.carrier_id = carrier_active.carrier_id;
+WHERE carrier.carrier_id = carrier_active.carrier_id
+    AND ((carrier_active.scope = 'default' AND carrier_active.value = 1) OR carrier_active.scope != 'default')
+ORDER BY carrier_active.scope
 SQL;
 
         return $this->connection->executeQuery($sql)->fetchAll(\PDO::FETCH_ASSOC);
     }
 
+    protected function setCarriers(array &$store, array $carriers): void
+    {
+        $website = $store['website_id'];
+
+        $resultCarriers = [];
+        foreach ($carriers as $carrier) {
+            $carrierId = $carrier['carrier_id'];
+
+            if ($carrier['scope'] === 'websites' && $carrier['scope_id'] !== $website) {
+                continue;
+            }
+
+            if ((int) $carrier['value'] === 1) {
+                $resultCarriers[$carrierId] = $carrier;
+            } else {
+                unset($resultCarriers[$carrierId]);
+            }
+        }
+
+        $store['carriers'] = \array_values($resultCarriers);
+    }
+
     protected function fetchPayments(): array
     {
         $sql = <<<SQL
-SELECT payment.*
+SELECT payment_active.*
 FROM (
-      SELECT
-             REPLACE(REPLACE(config.path, '/title', ''), 'payment/', '') AS payment_id,
+         SELECT
+             REPLACE(REPLACE(config.path, '/title', ''), 'payment/', '') AS payment_id
+         FROM {$this->tablePrefix}core_config_data config
+         WHERE path LIKE 'payment/%/title' AND scope = 'default'
+     ) AS payment,
+     (
+         SELECT
+             REPLACE(REPLACE(config.path, '/active', ''), 'payment/', '') AS payment_id,
              config.*
-      FROM {$this->tablePrefix}core_config_data config
-      WHERE path LIKE 'payment/%/title' AND scope = 'default'
-      ) AS payment,
-      (
-      SELECT
-             REPLACE(REPLACE(config.path, '/active', ''), 'payment/', '') AS payment_id
-      FROM {$this->tablePrefix}core_config_data config
-      WHERE path LIKE 'payment/%/active' AND scope = 'default' AND value = true
-      ) AS payment_active
-WHERE payment.payment_id = payment_active.payment_id;
+         FROM {$this->tablePrefix}core_config_data config
+         WHERE path LIKE 'payment/%/active'
+     ) AS payment_active
+WHERE payment.payment_id = payment_active.payment_id
+  AND ((payment_active.scope = 'default' AND payment_active.value = 1) OR payment_active.scope != 'default')
+ORDER BY payment_active.scope;
 SQL;
 
         return $this->connection->executeQuery($sql)->fetchAll(\PDO::FETCH_ASSOC);
+    }
+
+    protected function setPayments(array &$store, array $payments): void
+    {
+        $website = $store['website_id'];
+
+        $resultPayments = [];
+        foreach ($payments as $payment) {
+            $paymentId = $payment['payment_id'];
+
+            if ($payment['scope'] === 'websites' && $payment['scope_id'] !== $website) {
+                continue;
+            }
+
+            if ((int) $payment['value'] === 1) {
+                $resultPayments[$paymentId] = $payment;
+            } else {
+                unset($resultPayments[$paymentId]);
+            }
+        }
+
+        $store['payments'] = \array_values($resultPayments);
+    }
+
+    protected function fetchWebsiteConfig(array $websiteIds): array
+    {
+        $query = $this->connection->createQueryBuilder();
+
+        $query->from($this->tablePrefix . 'core_config_data', 'config');
+        $query->addSelect('scope_id AS website_id');
+        $query->addSelect('path');
+        $query->addSelect('value');
+
+        $query->andWhere('scope = \'websites\'');
+        $query->andWhere('scope_id IN (:websiteId)');
+        $query->andWhere('(path = \'currency/options/allow\' OR path = \'general/locale/code\' OR path = \'currency/options/default\' OR path = \'general/country/allow\' OR path = \'general/country/default\')');
+        $query->setParameter('websiteId', $websiteIds, Connection::PARAM_STR_ARRAY);
+
+        $query = $query->execute();
+        if (!($query instanceof ResultStatement)) {
+            return [];
+        }
+
+        $configs = $query->fetchAll(\PDO::FETCH_GROUP);
+
+        $returnConfig = [];
+        foreach ($configs as $key => $config) {
+            foreach ($config as $entry) {
+                if (!isset($entry['path'])) {
+                    continue;
+                }
+
+                $valueKey = null;
+                $value = $entry['value'];
+                switch ($entry['path']) {
+                    case 'general/locale/code':
+                        $valueKey = 'defaultLocale';
+                        $value = \str_replace('_', '-', $value);
+                        break;
+                    case 'currency/options/default':
+                        $valueKey = 'defaultCurrency';
+                        break;
+                    case 'currency/options/allow':
+                        $valueKey = 'allowedCurrencies';
+                        $value = \explode(',', $value);
+                        break;
+                    case 'general/country/allow':
+                        $valueKey = 'allowedCountries';
+                        $value = \explode(',', $value);
+                        break;
+                    case 'general/country/default':
+                        $valueKey = 'defaultCountry';
+                        break;
+                }
+
+                if ($valueKey === null) {
+                    continue;
+                }
+
+                $returnConfig[$key][$valueKey] = $value;
+            }
+        }
+
+        return $returnConfig;
     }
 
     protected function fetchStoreCurrencies(array $storeIds): array
@@ -335,5 +441,38 @@ SQL;
         $configurations['stores'] = $storeConfigs;
 
         return $configurations;
+    }
+
+    protected function setDefaultConfig(array &$store, array $defaults, array $websiteConfigs): void
+    {
+        $websiteId = $store['website_id'];
+
+        $store['currencies'] = $defaults['defaultAllowedCurrencies'];
+        if (isset($websiteConfigs[$websiteId]['allowedCurrencies'])) {
+            $store['currencies'] = $websiteConfigs[$websiteId]['allowedCurrencies'];
+        }
+
+        $store['defaultCurrency'] = $defaults['defaultCurrency'];
+        if (isset($websiteConfigs[$websiteId]['defaultCurrency'])) {
+            $store['defaultCurrency'] = $websiteConfigs[$websiteId]['defaultCurrency'];
+        }
+
+        if (isset($websiteConfigs[$websiteId]['defaultLocale'])) {
+            $store['locales'][] = $websiteConfigs[$websiteId]['defaultLocale'];
+            $store['defaultLocale'] = $websiteConfigs[$websiteId]['defaultLocale'];
+        } else {
+            $store['locales'][] = $defaults['defaultLocale'];
+            $store['defaultLocale'] = $defaults['defaultLocale'];
+        }
+
+        $store['countries'] = $defaults['defaultAllowedCountries'];
+        if (isset($websiteConfigs[$websiteId]['allowedCountries'])) {
+            $store['countries'] = $websiteConfigs[$websiteId]['allowedCountries'];
+        }
+
+        $store['defaultCountry'] = $defaults['defaultCountry'];
+        if (isset($websiteConfigs[$websiteId]['defaultCountry'])) {
+            $store['defaultCountry'] = $websiteConfigs[$websiteId]['defaultCountry'];
+        }
     }
 }